Link DOM elements with the game logic
Added a ScreenController function that is in charge of displaying the
the game state. First of all, a modal is opened when the page loads,
it takes the player's names input and display them on screen along with
the board. Every time a "cell" of the board is clicked it shows the
player token until a winning or tie condition is met. Finally it opens a
modal that shows the "winning state".

Also, added clearBoard functions to GameBoard and GameController, which
are in charge of "resetting" the game.